% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{as_flextable.binom_confint}
\alias{as_flextable.binom_confint}
\title{Convert \link{binom_confint} to \link[flextable]{flextable}}
\usage{
\method{as_flextable}{binom_confint}(x, ...)
}
\arguments{
\item{x}{a \link{binom_confint}}

\item{...}{..}
}
\description{
Convert \link{binom_confint} to \link[flextable]{flextable}
}
\keyword{internal}
