% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_array.R
\name{as_flextable.htest_array}
\alias{as_flextable.htest_array}
\title{\link{as_flextable.htest_array}}
\usage{
\method{as_flextable}{htest_array}(x, which = c("estimate", "p.value", "p.adjust"), ...)
}
\arguments{
\item{x}{..}

\item{which}{..}

\item{...}{..}
}
\description{
\link{as_flextable.htest_array}
}
\keyword{internal}
