% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{htest.tzh-package}
\alias{htest.tzh}
\alias{htest.tzh-package}
\title{htest.tzh: htest.tzh}
\description{
Tingting's extension to htest objects from package stats.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com}

}
\keyword{internal}
