% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{md_.binom_confint}
\alias{md_.binom_confint}
\title{md_.binom_confint}
\usage{
\method{md_}{binom_confint}(x, xnm, ...)
}
\arguments{
\item{x}{a \link{binom_confint}}

\item{xnm}{..}

\item{...}{..}
}
\description{
..
}
\examples{
list(
 'State Region' = state.region |> binom_confint() |> sort()
) |> fastmd::render_(file = 'binom_confint')
}
\keyword{internal}
