% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_array.R
\name{md_.htest_array}
\alias{md_.htest_array}
\title{Markdown Script of \link{htest_array}}
\usage{
\method{md_}{htest_array}(x, xnm, ...)
}
\arguments{
\item{x}{an \link{htest_array} object}

\item{xnm}{\link[base]{character} scalar, call of \code{x}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=md_.htest_array]{md_.htest_array()}} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Markdown Script of \link{htest_array}
}
\keyword{internal}
