% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_array.R
\name{p_adjust_.htest_array}
\alias{p_adjust_.htest_array}
\title{Adjusted \eqn{p}-values for \link{htest_array}}
\usage{
\method{p_adjust_}{htest_array}(x)
}
\arguments{
\item{x}{a \link{htest_array} object}
}
\description{
Adjusted \eqn{p}-values for \link{htest_array}
}
\keyword{internal}
