% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_array.R
\name{print.htest_array}
\alias{print.htest_array}
\title{print \code{htest_array}}
\usage{
\method{print}{htest_array}(x, ...)
}
\arguments{
\item{x}{..}
}
\description{
print \code{htest_array}
}
\keyword{internal}
