% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{sort.binom_confint}
\alias{sort.binom_confint}
\title{Sort \link{binom_confint}}
\usage{
\method{sort}{binom_confint}(x, decreasing = TRUE, ...)
}
\arguments{
\item{x}{\link{binom_confint}}

\item{decreasing, ...}{parameters of function \link[base]{order}}
}
\description{
Sort a \link{binom_confint} by count \code{x}, if the sample size \code{n} is the same.
}
\examples{
penguins |> is.na() |> binom_confint()
penguins |> is.na() |> binom_confint() |> sort()

}
\keyword{internal}
