% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{subset.binom_confint}
\alias{subset.binom_confint}
\title{Subset \link{binom_confint}}
\usage{
\method{subset}{binom_confint}(x, subset, ...)
}
\arguments{
\item{x}{\link{binom_confint}}

\item{subset}{\link[base]{language}}

\item{...}{additional parameters, currently of no use}
}
\description{
Subset a \link{binom_confint}.
}
\examples{
(cint = penguins |> is.na() |> binom_confint())
cint |> 
  subset(subset = x > 0L) |>
  sort()

}
\keyword{internal}
