% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{viewBinomCI}
\alias{viewBinomCI}
\title{View Binomial Confidence Interval}
\usage{
viewBinomCI(x)
}
\arguments{
\item{x}{a \link[base]{logical} \link[base]{matrix}}
}
\description{
..
}
\examples{
swiss |> is.na() |> viewBinomCI() # no missing
airquality |> is.na() |> viewBinomCI()
}
\keyword{internal}
