% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_confint.R
\name{binom_confint}
\alias{binom_confint}
\title{Clopper-Pearson Exact Binomial Confidence Interval}
\usage{
binom_confint(
  x,
  n,
  conf.level = 0.95,
  alternative = c("two.sided", "less", "greater"),
  ...
)
}
\arguments{
\item{x}{positive \link[base]{integer} scalar or \link[base]{vector}, counts}

\item{n}{positive \link[base]{integer} scalar or \link[base]{vector}, sample sizes \eqn{n}}

\item{conf.level, alternative, ...}{additional parameters of function \link[stats]{binom.test}}
}
\value{
Function \code{\link[=binom_confint]{binom_confint()}} returns an S3 class \code{'binom_confint'},
which is a \link[base]{matrix} with additional \link[base]{attributes}
\code{'conf.level'}, \code{'alternative'}, \code{'x'} and \code{'n'}.
}
\description{
Clopper-Pearson exact binomial confidence interval.
}
\examples{
library(flextable)
binom_confint(0:10, 10L) |> as_flextable()
binom_confint(0:10, 10L, alternative = 'less') |> as_flextable()
binom_confint(0:10, 10L, alternative = 'greater') |> as_flextable()
}
\keyword{internal}
