% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htest_array.R
\name{htest_array}
\alias{htest_array}
\alias{outer.cor.test}
\title{Array of \code{htest}}
\usage{
outer.cor.test(X, Y = X, ...)
}
\arguments{
\item{X, Y}{two \link[base]{matrix}es with same number of rows}

\item{...}{additional parameters of \link[stats]{cor.test}}
}
\value{
Function \code{\link[=outer.cor.test]{outer.cor.test()}} returns a \link{htest_array} object.
}
\description{
..
}
\details{
..
}
\note{
This is a factorial structure of \code{xc}-by-\code{yc}
\emph{not} a pairwise \emph{combination} of \code{x}.
}
\examples{
set.seed(100)
x = matrix(rnorm(50), ncol = 5, dimnames = list(NULL, LETTERS[1:5]))
y = matrix(rnorm(30), ncol = 3, dimnames = list(NULL, letters[1:3]))
m = outer.cor.test(x, y) # do no have print method yet
library(flextable.tzh)
library(flextable)
library(scales.tzh)
m$estimate |> as_flextable()
m$p.value |> label_pvalue_sym(add_p = TRUE)() |> as_flextable()
m |> p_adjust_.htest_array() |> label_pvalue_sym(add_p = TRUE)() |> as_flextable()

library(rmd.tzh); list(
 '`htest_array`' = m
) |> render_(file = 'htest_array')

}
\keyword{internal}
